move ServerDetails class to its own file
